feat: Added video Now with video support!  Watch your favorite meme videos with the MemeCataloger!  Video killed the bitmap star! (Kinda can't believe this took this long... wasn't THAT hard to build this part of it...)
